+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Show more information for runs in error state
+- Improve run stopping in case immediate run stop was not possible
 
 ## v1.0.1 - 2020-06-05
 ### Fixed